add proper env validation and update package exports
