+++ apps/api/.env.example
@@ -1,0 +1,24 @@
+VITE_BASE_URL="http://localhost:5183"
+APP_URL="http://localhost:5183"
+
+# Database
+DATABASE_URL=""
+
+# Auth
+BETTER_AUTH_SECRET=""
+BETTER_AUTH_URL="http://localhost:5183"
+
+GOOGLE_CLIENT_ID=""
+GOOGLE_CLIENT_SECRET=""
+
+# Redis
+REDIS_REST_URL="https://xyz.upstash.io"
+REDIS_REST_TOKEN="xyz"
+
+# Stripe
+STRIPE_SECRET_KEY="xyz"
+STRIPE_WEBHOOK_SECRET="xyz"
+STRIPE_PLATFORM_FEE_PERCENT="0"
+
+# OpenPanel
+VITE_OPENPANEL_CLIENT_ID="xyz"
