+++ apps/api/package.json
@@ -11,12 +11,12 @@
     "@hono/zod-validator": "^0.7.2",
     "@openads/auth": "workspace:*",
     "@openads/db": "workspace:*",
+    "@openads/redis": "workspace:*",
     "@openads/stripe": "workspace:*",
     "@openads/trpc": "workspace:*",
     "@openads/utils": "workspace:*",
     "@t3-oss/env-core": "^0.13.8",
     "@trpc/server": "^11.5.1",
-    "@upstash/redis": "^1.35.3",
     "hono": "^4.9.7",
     "superjson": "^2.2.2",
     "zod": "^4.1.8"
