+++ apps/api/src/services/auth.ts
