+++ apps/api/src/services/redis.ts
@@ -1,7 +1,7 @@
-import { Redis } from "@upstash/redis"
+import { createRedisClient } from "@openads/redis"
 import { env } from "~/env"
 
-export const redis = new Redis({
-  url: env.REDIS_REST_URL,
-  token: env.REDIS_REST_TOKEN,
+export const redis = createRedisClient({
+  REDIS_REST_URL: env.REDIS_REST_URL,
+  REDIS_REST_TOKEN: env.REDIS_REST_TOKEN,
 })
