+++ apps/api/src/services/stripe.ts
@@ -1,4 +1,4 @@
-import { createStripeClient, createPaymentIntent as createPaymentIntentCore } from "@openads/stripe"
+import { createPaymentIntent as createPaymentIntentCore, createStripeClient } from "@openads/stripe"
 import { env } from "~/env"
 
 export const stripe = createStripeClient({
