+++ biome.json
@@ -1,5 +1,5 @@
 {
-  "$schema": "https://biomejs.dev/schemas/2.1.2/schema.json",
+  "$schema": "https://biomejs.dev/schemas/2.2.4/schema.json",
   "assist": { "actions": { "source": { "organizeImports": "on" } } },
   "formatter": {
     "enabled": true,
@@ -35,9 +35,6 @@
     "enabled": true,
     "rules": {
       "recommended": true,
-      "nursery": {
-        "useUniqueElementIds": "off"
-      },
       "suspicious": {
         "noExplicitAny": "off",
         "noArrayIndexKey": "off",
