+++ packages/db/src/env.ts
@@ -1,0 +1,10 @@
+import { createEnv } from "@t3-oss/env-core";
+import { z } from "zod";
+
+export const env = createEnv({
+  server: {
+    DATABASE_URL: z.string(),
+  },
+
+  runtimeEnv: process.env,
+})
