+++ packages/db/src/lib/extensions.ts
@@ -1,5 +1,5 @@
 import { Prisma, WorkspaceMemberRole } from "@prisma/client"
-import { generateId } from "./lib/ids"
+import { generateId } from "./utils"
 
 type RecordData = Record<string, any>
