+++ packages/db/src/schema.ts
@@ -1,0 +1,5 @@
+export * from "./schema/booking"
+export * from "./schema/field"
+export * from "./schema/shared"
+export * from "./schema/spot"
+export * from "./schema/workspace"
