--- packages/db/src/schema/index.ts
@@ -1,5 +1,0 @@
-export * from "../lib/utils"
-export * from "./booking"
-export * from "./field"
-export * from "./spot"
-export * from "./workspace"
