+++ packages/db/src/schema/workspace.ts
@@ -1,13 +1,10 @@
 import { z } from "zod"
-import { isAllowedSlug, slugSchema } from "../lib/utils"
+import { slugSchema } from "./shared"
 
 export const workspaceSchema = z.object({
   // id: z.string().optional(),
   name: z.string().trim().min(3, { message: "Name is too short" }),
-  slug: slugSchema
-    .min(3, { message: "Slug is too short" })
-    .refine(isAllowedSlug, { message: "This slug is reserved" })
-    .default(""),
+  slug: slugSchema.min(3, { message: "Slug is too short" }).default(""),
   websiteUrl: z.url().trim().min(1, { message: "URL is required" }),
 })
