+++ packages/redis/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "@openads/tsconfig/base.json",
+  "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "~/*": ["./src/*"]
+    }
+  },
+  "include": ["src/**/*"],
+  "exclude": ["node_modules", "dist"]
+}
