+++ packages/trpc/package.json
@@ -4,20 +4,14 @@
   "type": "module",
   "exports": {
     ".": "./src/index.ts",
-    "./router": "./src/router.ts",
-    "./routers/auth": "./src/routers/auth.ts",
-    "./routers/booking": "./src/routers/booking.ts",
-    "./routers/field": "./src/routers/field.ts",
-    "./routers/onboarding": "./src/routers/onboarding.ts",
-    "./routers/spot": "./src/routers/spot.ts",
-    "./routers/stripe": "./src/routers/stripe.ts",
-    "./routers/user": "./src/routers/user.ts",
-    "./routers/workspace": "./src/routers/workspace.ts"
+    "./*": "./src/*.ts"
   },
   "dependencies": {
     "@openads/auth": "workspace:*",
     "@openads/db": "workspace:*",
+    "@openads/redis": "workspace:*",
     "@openads/stripe": "workspace:*",
+    "@t3-oss/env-core": "^0.13.8",
     "@trpc/server": "^11.5.1",
     "superjson": "^2.2.2",
     "zod": "^4.1.8"
